Move scopes into the checker since its not used elsewhere
